+++ pkg/metrics/metrics.go
@@ -1,0 +1,17 @@
+package metrics
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+var Deploys = prometheus.NewCounterVec(
+	prometheus.CounterOpts{
+		Name: "eo_deploys_total",
+		Help: "Deploy requests received from clients.",
+	},
+	[]string{"status"},
+)
+
+func init() {
+	prometheus.MustRegister(Deploys)
+}